chore: move code from main class to new files and adding documentation
